getArchetypeCatalog now uses actual locaRepository
