+++ archetype-common/src/main/java/org/apache/maven/archetype/ArchetypeManager.java
@@ -58,38 +58,33 @@ public interface ArchetypeManager
     ArchetypeCatalog getInternalCatalog();
 
     /**
-     * Gives the catalog of archetypes located in <code>${user.home}/.m2/repository/archetype-catalog.xml</code>.
-     * 
-     * @return the catalog.
-     */
-    ArchetypeCatalog getDefaultLocalCatalog();
-
-    /**
      * Gives the catalog of archetypes located in the given path.
      * if path is a file, it used as is.
      * if path is a directory, archetype-catalog.xml is appended to it.
      * 
-     * @param path the catalog file path or directory containing the catalog file.
+     * @param buildingRequest the catalog file path or directory containing the catalog file.
      * @return the catalog.
      */
-    ArchetypeCatalog getLocalCatalog( String path );
+    ArchetypeCatalog getLocalCatalog( ProjectBuildingRequest buildingRequest );
 
     /**
      * Gives the catalog of archetypes located at
      * <code>http://repo.maven.apache.org/maven2/archetype-catalog.xml</code>.
+     * @param buildingRequest TODO
      * 
      * @return the catalog.
      */
-    ArchetypeCatalog getRemoteCatalog();
+    ArchetypeCatalog getRemoteCatalog( ProjectBuildingRequest buildingRequest );
 
     /**
      * Gives the catalog of archetypes located at the given url.
      * if the url doesn't define a catalog, then <code>'archetype-catalog.xml'</code> is appended to it for search.
-     * 
+     * @param buildingRequest TODO
      * @param url the catalog url or base url containing the catalog file.
+     * 
      * @return the catalog.
      */
-    ArchetypeCatalog getRemoteCatalog( String url );
+    ArchetypeCatalog getRemoteCatalog( ProjectBuildingRequest buildingRequest, String url );
 
     /**
      * Creates a jar file for an archetype.
