+++ archetype-common/src/test/java/org/apache/maven/archetype/ArchetypeCatalogsTest.java
@@ -21,9 +21,13 @@ import org.apache.commons.io.IOUtils;
 import org.apache.maven.archetype.catalog.Archetype;
 import org.apache.maven.archetype.catalog.ArchetypeCatalog;
 import org.apache.maven.archetype.catalog.io.xpp3.ArchetypeCatalogXpp3Writer;
+import org.apache.maven.project.DefaultProjectBuildingRequest;
+import org.apache.maven.project.ProjectBuildingRequest;
+import org.apache.maven.repository.internal.MavenRepositorySystemSession;
 import org.codehaus.plexus.PlexusTestCase;
 import org.mortbay.jetty.Server;
 import org.mortbay.jetty.webapp.WebAppContext;
+import org.sonatype.aether.impl.internal.SimpleLocalRepositoryManager;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -38,9 +42,11 @@ public class ArchetypeCatalogsTest
     public void testRemoteCatalog()
         throws Exception
     {
-        ArchetypeManager archetype = (ArchetypeManager) lookup( ArchetypeManager.class.getName() );
+        ArchetypeManager archetype = lookup( ArchetypeManager.class );
 
-        ArchetypeCatalog result = archetype.getRemoteCatalog( "http://localhost:" + port + "/repo/" );
+        ProjectBuildingRequest buildingRequest = new DefaultProjectBuildingRequest();
+        
+        ArchetypeCatalog result = archetype.getRemoteCatalog( buildingRequest, "http://localhost:" + port + "/repo/" );
 
         assertEquals( 1, result.getArchetypes().size() );
         assertEquals( "groupId", result.getArchetypes().get( 0 ).getGroupId() );
@@ -52,11 +58,15 @@ public class ArchetypeCatalogsTest
     public void testLocalCatalog()
         throws Exception
     {
-        ArchetypeManager archetype = (ArchetypeManager) lookup( ArchetypeManager.class.getName() );
+        ArchetypeManager archetype = lookup( ArchetypeManager.class );
 
-        ArchetypeCatalog result =
-            archetype.getLocalCatalog( getTestFile( "target/test-classes/repositories/test-catalog" ).
-                getAbsolutePath() );
+        ProjectBuildingRequest buildingRequest = new DefaultProjectBuildingRequest();
+        MavenRepositorySystemSession repositorySession = new MavenRepositorySystemSession();
+        repositorySession.setLocalRepositoryManager( new SimpleLocalRepositoryManager( getTestFile( "target/test-classes/repositories/test-catalog" ) ) );
+        buildingRequest.setRepositorySession( repositorySession );
+        
+        
+        ArchetypeCatalog result = archetype.getLocalCatalog( buildingRequest );
 
         assertEquals( 1, result.getArchetypes().size() );
         assertEquals( "groupId", result.getArchetypes().get( 0 ).getGroupId() );
